+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Li 2017 naproxen ([doi:10.1124/dmd.116.074500](https://doi.org/10.1124/dmd.116.074500)) -- male and female Lewis rats with collagen-induced arthritis and healthy controls.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.